# ⚡ `RocksDBStore` Concat optimization