Tested with 1,000,000 byte string pairs (or triples) of length `WordSize` below. Obviously as words get longer, the proportion of LINQ overhead decreases. However, for the most part, we are only concatenating small byte strings for most, if not all, cases.

BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19045.3086)
11th Gen Intel Core i7-1185G7 3.00GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.203
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2


|            Method | WordSize |        Mean |     Error |     StdDev |      Median |        Gen0 |  Allocated |
|------------------ |--------- |------------:|----------:|-----------:|------------:|------------:|-----------:|
|   WithConcatPairs |       10 |   243.32 ms |  4.791 ms |  10.617 ms |   244.86 ms |  25333.3333 |  160.22 MB |
|     WithCopyPairs |       10 |    94.13 ms |  1.872 ms |   5.432 ms |    93.98 ms |   7666.6667 |   53.41 MB |
| WithConcatTriples |       10 |   314.55 ms |  9.623 ms |  28.223 ms |   313.11 ms |  28000.0000 |  175.48 MB |
|   WithCopyTriples |       10 |   102.91 ms |  2.046 ms |   4.180 ms |   103.41 ms |   9000.0000 |   61.04 MB |
|   WithConcatPairs |      100 |   432.83 ms |  9.997 ms |  29.161 ms |   431.51 ms |  53000.0000 |  328.06 MB |
|     WithCopyPairs |      100 |   235.76 ms |  4.667 ms |   9.639 ms |   235.96 ms |  35666.6667 |  221.25 MB |
| WithConcatTriples |      100 |   626.57 ms | 13.426 ms |  39.377 ms |   627.25 ms |  71000.0000 |  434.88 MB |
|   WithCopyTriples |      100 |   312.88 ms | 10.005 ms |  28.705 ms |   304.96 ms |  52000.0000 |  320.43 MB |
|   WithConcatPairs |     1000 | 1,339.92 ms | 85.190 ms | 251.185 ms | 1,379.64 ms | 340000.0000 | 2044.68 MB |
|     WithCopyPairs |     1000 | 1,002.98 ms | 67.109 ms | 195.760 ms |   963.25 ms | 322000.0000 | 1937.87 MB |
| WithConcatTriples |     1000 | 2,879.15 ms | 53.529 ms |  50.071 ms | 2,876.16 ms | 502000.0000 | 3005.99 MB |
|   WithCopyTriples |     1000 | 2,209.61 ms | 42.739 ms |  59.914 ms | 2,209.95 ms | 483000.0000 | 2891.54 MB |